Fix: use jsx-dev-runtime when development (#6883)
* fix: use jsx-dev-runtime when development

* chore: import path
