+++ .changeset/few-meals-listen.md
@@ -1,0 +1,5 @@
+---
+'@ice/jsx-runtime': minor
+---
+
+fix: use jsx-dev-runtime when development
